Refactor link transformation in link utils
