+++ apps/web/app/api/links/info/route.ts
@@ -1,8 +1,8 @@
 import { DubApiError } from "@/lib/api/errors";
+import { transformLink } from "@/lib/api/links";
 import { withAuth } from "@/lib/auth";
 import prisma from "@/lib/prisma";
 import { getLinkInfoQuerySchema } from "@/lib/zod/schemas/links";
-import { linkConstructor } from "@dub/utils";
 import { NextResponse } from "next/server";
 
 // GET /api/links/info – get the info for a link
@@ -18,6 +18,17 @@ export const GET = withAuth(async ({ headers, searchParams }) => {
     },
     include: {
       user: true,
+      tags: {
+        select: {
+          tag: {
+            select: {
+              id: true,
+              name: true,
+              color: true,
+            },
+          },
+        },
+      },
     },
   });
 
@@ -28,37 +39,7 @@ export const GET = withAuth(async ({ headers, searchParams }) => {
     });
   }
 
-  const tags = await prisma.tag.findMany({
-    where: {
-      linksNew: {
-        some: {
-          linkId: link.id,
-        },
-      },
-    },
-    select: {
-      id: true,
-      name: true,
-      color: true,
-    },
+  return NextResponse.json(transformLink(link), {
+    headers,
   });
-
-  const shortLink = linkConstructor({
-    domain: link.domain,
-    key: link.key,
-  });
-
-  return NextResponse.json(
-    {
-      ...link,
-      shortLink,
-      tagId: tags?.[0]?.id ?? null, // backwards compatibility
-      tags,
-      qrCode: `https://api.dub.co/qr?url=${shortLink}`,
-      workspaceId: `ws_${link.projectId}`,
-    },
-    {
-      headers,
-    },
-  );
 });
