+++ apps/web/lib/api/links/utils.ts
@@ -7,9 +7,15 @@ import {
   SHORT_DOMAIN,
   getApexDomain,
   isDubDomain,
+  linkConstructor,
   log,
   validKeyRegex,
 } from "@dub/utils";
+import { Link, Tag } from "@prisma/client";
+
+type LinkWithTags = Link & {
+  tags: { tag: Pick<Tag, "id" | "name" | "color"> }[];
+};
 
 export function combineTagIds({
   tagId,
@@ -117,3 +123,22 @@ export async function dubLinkChecks(link: SimpleLinkProps) {
   }
   return null;
 }
+
+// Transform link with additional properties
+export const transformLink = (link: LinkWithTags) => {
+  const tags = link.tags.map(({ tag }) => tag);
+
+  const shortLink = linkConstructor({
+    domain: link.domain,
+    key: link.key,
+  });
+
+  return {
+    ...link,
+    shortLink,
+    tagId: tags?.[0]?.id ?? null, // backwards compatibility
+    tags,
+    qrCode: `https://api.dub.co/qr?url=${shortLink}`,
+    workspaceId: `ws_${link.projectId}`,
+  };
+};
